error handling and some refactors
