+++ src/interfaces/Answer.ts
@@ -1,0 +1,7 @@
+interface Answer {
+    letter: string;
+    text: string;
+    is_correct: boolean;
+}
+
+export default Answer;
