+++ src/styles/GameModes.css
@@ -15,6 +15,7 @@
     color: white;
     background-color: rgba(28, 94, 164, 0.5);
     font-family: 'Chalkduster', serif;
+    text-decoration: none;
 }
 
 .game-mode-button:hover {
